ENH: Compute Avg Orientations now can use von-Mises Fisher or Watson sampling (#1577)
* ENH: Compute Avg Orientations now can use von-Mises Fisher or Watson sampling.
These algorithms are based on algorithms from the EMsoft software library.
* BUG: Fix SIMPLJson conversion test
* BUG: Fix phaseIndex calculation bug.

+++ src/Plugins/OrientationAnalysis/src/OrientationAnalysis/Filters/Algorithms/ComputeFZQuaternions.cpp
@@ -153,8 +153,8 @@ Result<> ComputeFZQuaternions::operator()()
     }
     else if(maskArray->getDataType() == DataType::uint8)
     {
-      UInt32Array* goodVoxelsArray = m_DataStructure.getDataAs<UInt32Array>(m_InputValues->MaskArrayPath);
-      dataAlg.execute(::GenerateFZQuatsImpl<UInt32Array>(quatArray, phaseArray, xtalArray, numPhases, goodVoxelsArray, fzQuatArray, m_ShouldCancel, warningCount));
+      UInt8Array* goodVoxelsArray = m_DataStructure.getDataAs<UInt8Array>(m_InputValues->MaskArrayPath);
+      dataAlg.execute(::GenerateFZQuatsImpl<UInt8Array>(quatArray, phaseArray, xtalArray, numPhases, goodVoxelsArray, fzQuatArray, m_ShouldCancel, warningCount));
     }
     else if(maskArray->getDataType() == DataType::int8)
     {
